+++ lib/src/model/import.dart
@@ -1,0 +1,5 @@
+class Import {
+  final String package;
+
+  Import(this.package);
+}