+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.2
 
+- Feature: custom imports defined on xml level (import node)
 - Feature: type now can be given in xml node name, no need to use field node
 - Feature: add comment to generated field (attribute comment)
 - Improvement: More helper functions (duration)